Database

Description

Module to Manage the Creation of the database and any operations performed on the database.

Documentation

initialiseDBPlanets :: IO Connection

Connects to the database and creates a Planets table

This function takes no arguments

This function returns an IO with a HDBC Connection

initialiseDBPeople :: IO Connection

Connects to the database and creates a People table

This function takes no arguments

This function returns an IO with a HDBC Connection

initialiseDBSpecies :: IO Connection

Connects to the database and creates a Species table

This function takes no arguments

This function returns an IO with a HDBC Connection

initialiseDBFilms :: IO Connection

Connects to the database and creates a Films table

This function takes no arguments

This function returns an IO with a HDBC Connection

savePlanets

Arguments

:: [Planet]
takes an array containing type Planet

-> Connection
takes an HDBC Connection

-> IO ()
Returns an empty IO

Saves values to the Planets table

It takes two arguments and returns an empty IO

savePeople

Arguments

:: [Person]
takes an array containing type Person

-> Connection
takes an HDBC Connection

-> IO ()
Returns an empty IO

Saves values to the Person table

It takes two arguments and returns an empty IO

saveSpecies

Arguments

:: [Species]
takes an array containing type Species

-> Connection
takes an HDBC Connection

-> IO ()
Returns an empty IO

Saves values to the Species table

It takes two arguments and returns an empty IO

saveFilms

Arguments

:: [Film]
takes an array containing type Film

-> Connection
takes an HDBC Connection

-> IO ()
Returns an empty IO

Saves values to the Film table

It takes two arguments and returns an empty IO

convertUnkToNothing

Arguments

:: String
Takes the String value returned from the API

-> Maybe String
Returns either Nothing (if the value is "unknown"), otherwise returns the value itself

Converts values of "unknown" to type Nothing

It should only be used if the JSON value is sometimes "unknown" - do not use for Null values

It takes a single argument

extractID

Arguments

:: Maybe [Char]
Takes a URL of type Maybe [Char]

-> Maybe [Char]
returns the ID value as type [Char] or Nothing (if Nothing is passed)

Gets the ID from any Star Wars API Url

It takes a single argument

planetToSqlValues

Arguments

:: Planet
Takes a Record of type Planet

-> [SqlValue]
returns an array of SQL Values

Transforms Planet values to SQL

It takes a single argument Record of type Planet and Returns an Array of SQL Values

prepareInsertPlanetSmt

Arguments

:: Connection
Takes a HDBC Connection

-> IO Statement
returns an IO with the prepared HDBC Statement

Prepares a Posgres statement that inserts values into the Planets table

It takes one argument and returns an IO with a prepared HDBC Statement

personToSqlValues

Arguments

:: Person
Takes a Record of type Person

-> [SqlValue]
returns an array of SQL Values

Transforms Person values to SQL

It takes a single argument Record of type Person and Returns an Array of SQL Values

prepareInsertPeopleSmt

Arguments

:: Connection
Takes a HDBC Connection

-> IO Statement
returns an IO with the prepared HDBC Statement

Prepares a Posgres statement that inserts values into the People table

It takes one argument and returns an IO with a prepared HDBC Statement

speciesToSqlValues

Arguments

:: Species
Takes a Record of type Species

-> [SqlValue]
returns an array of SQL Values

Transforms Species values to SQL

It takes a single argument Record of type Species and Returns an Array of SQL Values

prepareInsertSpeciesSmt

Arguments

:: Connection
Takes a HDBC Connection

-> IO Statement
returns an IO with the prepared HDBC Statement

Prepares a Posgres statement that inserts values into the Species table

It takes one argument and returns an IO with a prepared HDBC Statement

filmToSqlValues

Arguments

:: Film
Takes a Record of type Film

-> [SqlValue]
returns an array of SQL Values

Transforms Film values to SQL

It takes a single argument Record of type Film and Returns an Array of SQL Values

prepareInsertFilmSmt

Arguments

:: Connection
Takes a HDBC Connection

-> IO Statement
returns an IO with the prepared HDBC Statement

Prepares a Posgres statement that inserts values into the Film table

It takes one argument and returns an IO with a prepared HDBC Statement