Parse

Description

Module which helps parse JSON results

Documentation

parse

Arguments

:: ByteString
It takes one argument a ByteString

-> Either String AllResults
returns either an error String, or a Record of type AllResults

Decodes the JSON returned by the root API result

data AllResults

The AllResults constructor defines the types of the root API results

It is a derivation of the Show and Generic instances