ParseFilms

Description

Module which helps parse JSON results for the type Film

Documentation

parseFilms

Arguments

:: ByteString
It takes one argument a ByteString

-> Either String FilmResults
returns either an error String, or a Record of type FilmResults

Decodes the JSON returned by the films API result

data FilmResults

The FilmResults constructor defines the types of the Film API Results It is a derivation of the Show and Generic instances

data Film

The Film constructor defines the types of the Film Record

It is a derivation of the Show and Generic instances