ParsePeople

Description

Module which helps parse JSON results for the type People

Documentation

parsePeople

Arguments

:: ByteString
It takes one argument a ByteString

-> Either String PeopleResults
returns either an error String, or a Record of type PeopleResults

Decodes the JSON returned by the people API result

data PeopleResults

The PeopleResults constructor defines the types of the People API Results It is a derivation of the Show and Generic instances

data Person

The Person constructor defines the types of the Person Record It is a derivation of the Show and Generic instances