ParsePlanets

Description

Module which helps parse JSON results for the type Planets

Documentation

parsePlanets

Arguments

:: ByteString
It takes one argument a ByteString

-> Either String PlanetResults
returns either an error String, or a Record of type PlanetResults

Decodes the JSON returned by the Planets API result

data PlanetResults

The PlanetResults constructor defines the types of the Planet API Results It is a derivation of the Show and Generic instances

data Planet

The Planet constructor defines the types of the Planet Record It is a derivation of the Show and Generic instances