ParseSpecies

Description

Module which helps parse JSON results for the type Species

Documentation

parseSpecies

Arguments

:: ByteString
It takes one argument a ByteString

-> Either String SpeciesResults
returns either an error String, or a Record of type SpeciesResults

Decodes the JSON returned by the Species API result

data SpeciesResults

The SpeciesResults constructor defines the types of the Species API Results It is a derivation of the Show and Generic instances

data Species

The Species constructor defines the types of the Species Record It is a derivation of the Show and Generic instances