ProcessJson

Description

Module which helps process JSON results

Documentation

processJson

Arguments

:: String
Takes the URL to process

-> String
Takes the directory of the API e.g. "planets"

-> IO ()
returns an empty IO

Processes the API result for a given URL and directory This function takes two arguments

savePlanetJson

Arguments

:: PlanetResults
takes one argument the result from the Planets API directory

-> IO ()
returns an Empty IO

Saves Planet JSON results to the database

savePeopleJson

Arguments

:: PeopleResults
takes one argument the result from the People API directory

-> IO ()
returns an Empty IO

Saves People JSON results to the database

saveFilmsJson

Arguments

:: FilmResults
takes one argument the result from the Film API directory

-> IO ()
returns an Empty IO

Saves Film JSON results to the database

saveSpeciesJson

Arguments

:: SpeciesResults
takes one argument the result from the Species API directory

-> IO ()
returns an Empty IO

Saves Species JSON results to the database

catchErr :: Show a => (t1 -> Either a t2) -> t1 -> (t2 -> IO ()) -> IO ()

Catches any error while parsing a JSON If an error occurs, the error is printed. Otherwise the result is returned

It takes three arguments:

the parsing function to perform on a given JSON

a JSON

the save function to perform if the parse is successful

It returns an empty IO

parseDirectories

Arguments

:: ByteString
takes one argument the returned JSON of the root directory

-> IO ()
returns an empty IO

Parse Directories is performed on the API's root result.

It processes the JSONs for each directory within the root (planets, people, species, films)