Queries

Description

Module which contains the queries

Documentation

dropAllTables

Arguments

:: IO ()
Performs an IO action

Drops all tables from the database, ensures tables are always clean when re-running code

queryDeletePeople

Arguments

:: IO ()
Performs an IO action

Query to delete all people from a given planet

queryInsertToFilms

Arguments

:: String

-> String

-> String

-> String

-> String

-> String
Takes 6 strings as inputs

-> IO ()
Performs an IO action

Query to insert a new row into the films table

queryFilms

Arguments

:: IO ()
Performs an IO action

Query to retrive title and producer from the films table

queryHomeworlds

Arguments

:: IO ()
Performs an IO action

Query to find people and their homeworlds

queryPlanetWithHighestGround

Arguments

:: IO ()
Performs an IO action

Query to find and return the planet with the highest ground